GP2-1616: Move some reusable abstract classes into core ahead of more hybrid page porting
The Snippet-based approach to editable content for Django views was added for a view in `contact` but the base classes should be somewhere in core for other apps to leverage the same pattern.

This changeset does that, with no test regressions.

Further changes may follow.
