+++ cms_extras/modeladmin.py
@@ -3,9 +3,14 @@ import csv
 from django.utils.html import format_html_join, strip_tags
 from wagtail.admin.views.mixins import Echo
 from wagtail.contrib.modeladmin.helpers import ButtonHelper
-from wagtail.contrib.modeladmin.options import ModelAdmin, modeladmin_register
+from wagtail.contrib.modeladmin.options import (
+    ModelAdmin,
+    ModelAdminGroup,
+    modeladmin_register,
+)
 from wagtail.contrib.modeladmin.views import IndexView
 
+from contact.modeladmin import ContactSuccessSnippetAdmin
 from core.models import CaseStudy
 
 
@@ -195,3 +200,16 @@ class CaseStudyAdmin(ModelAdmin):
 
 
 modeladmin_register(CaseStudyAdmin)
+
+
+class NonCMSContentGroup(ModelAdminGroup):
+    """All of the snippets used for hybrid-content pages
+    should be registered as part of this group"""
+
+    menu_label = 'Non-page content'
+    menu_icon = 'folder-open-inverse'  # change as required
+    menu_order = 200
+    items = (ContactSuccessSnippetAdmin,)
+
+
+modeladmin_register(NonCMSContentGroup)
